Adapt customer_free_ref to postfinance
