+++ ebill_postfinance_customer_free_ref/__manifest__.py
@@ -2,12 +2,12 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
-    "name": "eBill Paynet Customer Free Reference",
-    "summary": "Glue module: ebill_paynet and sale_order_customer_free_ref",
-    "version": "14.0.1.0.1",
+    "name": "eBill Postfinance Customer Free Reference",
+    "summary": "Glue module: ebill_postfinance and sale_order_customer_free_ref",
+    "version": "14.0.1.0.0",
     "license": "AGPL-3",
-    "author": "Camptocamp SA,Odoo Community Association (OCA)",
+    "author": "Camptocamp,Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/l10n-switzerland",
-    "depends": ["ebill_paynet", "sale_order_customer_free_ref"],
+    "depends": ["ebill_postfinance", "sale_order_customer_free_ref"],
     "auto_install": True,
 }
