--- ebill_postfinance_customer_free_ref/i18n/ebill_paynet_customer_free_ref.pot
@@ -1,42 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* ebill_paynet_customer_free_ref
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: ebill_paynet_customer_free_ref
-#: model:ir.model.fields,field_description:ebill_paynet_customer_free_ref.field_account_move__display_name
-#: model:ir.model.fields,field_description:ebill_paynet_customer_free_ref.field_sale_order__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: ebill_paynet_customer_free_ref
-#: model:ir.model.fields,field_description:ebill_paynet_customer_free_ref.field_account_move__id
-#: model:ir.model.fields,field_description:ebill_paynet_customer_free_ref.field_sale_order__id
-msgid "ID"
-msgstr ""
-
-#. module: ebill_paynet_customer_free_ref
-#: model:ir.model,name:ebill_paynet_customer_free_ref.model_account_move
-msgid "Journal Entry"
-msgstr ""
-
-#. module: ebill_paynet_customer_free_ref
-#: model:ir.model.fields,field_description:ebill_paynet_customer_free_ref.field_account_move____last_update
-#: model:ir.model.fields,field_description:ebill_paynet_customer_free_ref.field_sale_order____last_update
-msgid "Last Modified on"
-msgstr ""
-
-#. module: ebill_paynet_customer_free_ref
-#: model:ir.model,name:ebill_paynet_customer_free_ref.model_sale_order
-msgid "Sales Order"
-msgstr ""
