+++ ebill_postfinance_customer_free_ref/models/sale_order.py
@@ -1,4 +1,4 @@
-# Copyright 2021 Camptocamp SA
+# Copyright 2021-2022 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import api, models
@@ -9,6 +9,6 @@ class SaleOrder(models.Model):
     _inherit = "sale.order"
 
     @api.depends("customer_order_number")
-    def _compute_paynet_client_order_ref(self):
+    def _compute_postfinance_ebill_client_order_ref(self):
         for order in self:
-            order.paynet_client_order_ref = order.customer_order_number
+            order.postfinance_ebill_client_order_ref = order.customer_order_number
