+++ ebill_postfinance_customer_free_ref/readme/CONTRIBUTORS.rst
@@ -1,3 +1,1 @@
 * Thierry Ducrest <thierry.ducrest@camptocamp.com>
-* `Trobz <https://trobz.com>`_:
-    * Khoi Vo <khoivha@trobz.com>
