--- ebill_postfinance_customer_free_ref/readme/CREDITS.rst
@@ -1,1 +1,0 @@
-The migration of this module from 13.0 to 14.0 was financially supported by Camptocamp
